mspr3-deployment/shared/docker/etl/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
psycopg2-binary==2.9.9
requests==2.31.0
python-multipart==0.0.6
